+++ scripts/publish.sh
@@ -1,4 +1,6 @@
-npm i
+#!/usr/bin/env bash
+set -eo pipefail
+
 npm run build -- --minify
 cp package.json dist/package.json
 cp README.md dist/README.md